ref(replays): Move all the react hooks into one folder
