+++ static/app/components/replays/hooks/useFullscreen.tsx
