+++ static/app/components/replays/hooks/useRAF.tsx
