+++ static/app/components/replays/replayContext.tsx
@@ -2,11 +2,11 @@ import React, {useCallback, useContext, useEffect, useRef, useState} from 'react
 import {useTheme} from '@emotion/react';
 import {Replayer, ReplayerEvents} from 'rrweb';
 
+import useRAF from 'sentry/components/replays/hooks/useRAF';
 import type ReplayReader from 'sentry/utils/replays/replayReader';
 import usePrevious from 'sentry/utils/usePrevious';
 
 import HighlightReplayPlugin from './highlightReplayPlugin';
-import useRAF from './useRAF';
 
 type Dimensions = {height: number; width: number};
 type RootElem = null | HTMLDivElement;
