+++ static/app/components/replays/replayController.tsx
@@ -4,8 +4,8 @@ import styled from '@emotion/styled';
 import Button from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
 import CompactSelect from 'sentry/components/forms/compactSelect';
+import useFullscreen from 'sentry/components/replays/hooks/useFullscreen';
 import {useReplayContext} from 'sentry/components/replays/replayContext';
-import useFullscreen from 'sentry/components/replays/useFullscreen';
 import {IconArrow, IconPause, IconPlay, IconRefresh, IconResize} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import space from 'sentry/styles/space';
